+++ src/main.tsx
@@ -1,10 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { MsalProvider } from '@azure/msal-react'
+import { msalInstance } from './auth/msalInstance'
 import './index.css'
 import App from './app/App'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <MsalProvider instance={msalInstance}>
+      <App />
+    </MsalProvider>
   </StrictMode>,
 )